+++ src/main/java/com/cobber/fta/LTRandom.java
@@ -17,5 +17,5 @@ package com.cobber.fta;
 
 public interface LTRandom {
 	public String nextRandom();
-	public void seed(long seed);
+	public void seed(byte[] seed);
 }