+++ src/main/java/com/cobber/fta/plugins/LogicalTypeNameLastFirst.java
@@ -32,6 +32,7 @@ import com.cobber.fta.StringFacts;
 public class LogicalTypeNameLastFirst extends LogicalTypeInfinite {
 	public static final String SEMANTIC_TYPE = "NAME.LAST_FIRST";
 	public static final String REGEXP = "[- \\p{IsAlphabetic}]+, ?[- \\p{IsAlphabetic}]+";
+	private static final String BACKOUT = ".+";
 	private LogicalTypeCode logicalFirst;
 	private LogicalTypeCode logicalLast;
 
@@ -141,7 +142,21 @@ public class LogicalTypeNameLastFirst extends LogicalTypeInfinite {
 	@Override
 	public String isValidSet(String dataStreamName, long matchCount, long realSamples, StringFacts stringFacts,
 			Map<String, Long> cardinality, Map<String, Long> outliers) {
-		return getConfidence(matchCount, realSamples, dataStreamName) >= getThreshold()/100.0 ? null : ".+";
+
+		int minCardinality = 8;
+		int minSamples = 10;
+		if (getHeaderConfidence(dataStreamName) != 0) {
+			minCardinality = 3;
+			minSamples = 3;
+		}
+
+		if (cardinality.size() < minCardinality)
+			return BACKOUT;
+
+		if (realSamples < minSamples)
+			return BACKOUT;
+
+		return getConfidence(matchCount, realSamples, dataStreamName) >= getThreshold()/100.0 ? null : BACKOUT;
 	}
 
 	@Override